Complete refactor of static information and navigation menu
All static information has been moved to the cms and can now be queried using the StaticInfoAPI module. All pages containing static info has been moved to the /info folder, and is routed as /info/...
The navigation menu is changed to a more usable sidebar component, which contains all relevant routes to the various static pages.